src/requirements.txt
Flask==3.0.2
mecab-python3==1.0.8
unidic-lite==1.0.8
